refactor(check_progress): extract shared logic into common.py
